backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
boto3==1.29.3
torch==2.1.1
tensorflow==2.14.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
prometheus-client==0.19.0
aiofiles==23.2.1
gputil==1.4.0
bayesian-optimization==1.4.3
joblib==1.3.2
